requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
psycopg[binary]==3.2.3
alembic==1.14.0
pydantic-settings==2.7.0
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
pandas==2.2.3
httpx==0.28.1
tenacity==9.0.0
celery==5.4.0
redis==5.2.1
PyMuPDF==1.25.1
pdfplumber==0.11.5
python-docx==1.1.2
pytesseract==0.3.13
Pillow==11.1.0
email-validator==2.2.0
